src/emailservice/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==4.2.0
    # via google-auth
certifi==2020.12.5
    # via requests
chardet==3.0.4
    # via requests
google-api-core[grpc]==1.24.1
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-monitoring
    #   google-cloud-trace
    #   opencensus
google-api-python-client==1.12.8
    # via google-cloud-profiler
google-auth-httplib2==0.0.4
    # via
    #   google-api-python-client
    #   google-cloud-profiler
google-auth==1.24.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-profiler
google-cloud-core==1.5.0
    # via google-cloud-trace
google-cloud-monitoring==0.36.0
    # via opencensus-ext-stackdriver
google-cloud-profiler==1.1.2
    # via -r requirements.in
google-cloud-trace==0.24.0
    # via
    #   -r requirements.in
    #   opencensus-ext-stackdriver
googleapis-common-protos==1.52.0
    # via google-api-core
grpcio-health-checking==1.33.2
    # via -r requirements.in
grpcio==1.33.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   grpcio-health-checking
    #   opencensus-ext-grpc
httplib2==0.19.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==2.10
    # via requests
jinja2==2.11.3
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
opencensus-context==0.1.2
    # via opencensus
opencensus-ext-grpc==0.7.1
    # via -r requirements.in
opencensus-ext-stackdriver==0.7.3
    # via -r requirements.in
opencensus==0.7.11
    # via
    #   -r requirements.in
    #   opencensus-ext-grpc
    #   opencensus-ext-stackdriver
protobuf==3.13.0
    # via
    #   google-api-core
    #   google-cloud-profiler
    #   googleapis-common-protos
    #   grpcio-health-checking
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyparsing==2.4.7
    # via httplib2
python-json-logger==2.0.2
    # via -r requirements.in
pytz==2018.9
    # via google-api-core
requests==2.24.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-profiler
rsa==4.7
    # via google-auth
six==1.15.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-core
    #   grpcio
    #   protobuf
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.25.10
    # via requests
opentelemetry-instrumentation-grpc==0.24b0
opentelemetry-sdk==1.5.0
opentelemetry-api==1.5.0
opentelemetry-exporter-otlp==1.5.0
# The following packages are considered to be unsafe in a requirements file:
# setuptools
